requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
aiogram~=3.10.0
python-decouple~=3.8
asyncpg-lite~=0.3.1.3
openai
selenium